refactor: use `ensureHashAppIsInitialized` in hash app backend
