+++ packages/hash/integration/src/tests/model/knowledge/page.model.test.ts
@@ -1,5 +1,8 @@
 import { getRequiredEnv } from "@hashintel/hash-backend-utils/environment";
-import { createGraphClient } from "@hashintel/hash-api/src/graph";
+import {
+  createGraphClient,
+  ensureHashAppIsInitialized,
+} from "@hashintel/hash-api/src/graph";
 import { SYSTEM_TYPES } from "@hashintel/hash-api/src/graph/system-types";
 import {
   BlockModel,
@@ -9,7 +12,7 @@ import {
 } from "@hashintel/hash-api/src/model";
 import { Logger } from "@hashintel/hash-backend-utils/logger";
 import { TypeSystemInitializer } from "@blockprotocol/type-system";
-import { createTestUser, ensureHashAppIsInitialized } from "../../util";
+import { createTestUser } from "../../util";
 
 jest.setTimeout(60000);
