+++ packages/hash/integration/src/tests/model/knowledge/user.model.test.ts
@@ -1,5 +1,8 @@
 import { getRequiredEnv } from "@hashintel/hash-backend-utils/environment";
-import { createGraphClient } from "@hashintel/hash-api/src/graph";
+import {
+  createGraphClient,
+  ensureHashAppIsInitialized,
+} from "@hashintel/hash-api/src/graph";
 import { Logger } from "@hashintel/hash-backend-utils/logger";
 
 import { UserModel } from "@hashintel/hash-api/src/model";
@@ -9,11 +12,7 @@ import {
 } from "@hashintel/hash-api/src/auth/ory-kratos";
 import { systemOrgAccountId } from "@hashintel/hash-api/src/graph/system-org";
 import { TypeSystemInitializer } from "@blockprotocol/type-system";
-import {
-  createTestOrg,
-  ensureHashAppIsInitialized,
-  generateRandomShortname,
-} from "../../util";
+import { createTestOrg, generateRandomShortname } from "../../util";
 
 jest.setTimeout(60000);
